requirements.txt
Django==5.0.2
Flask==3.0.2
Pillow==10.2.0
python-dotenv==1.0.1
requests==2.31.0
django-cors-headers==4.3.1
Werkzeug==3.0.1
gunicorn==21.2.0
whitenoise==6.6.0
django-tailwind==3.8.0 